Allowing repeated initialisation
I don't see a reason why this should not work, but obviously I don't really see the whole picture, so please review. We need this functionality in our app (iOS allows this). I Also fixed a few places where Context was being unsafely cast to Application. It is still not completely safe, but safer.